Use naive offset approach for turbo.cpp
The previous attempt does not work because the distance() method for
C++ sets is O(n). Modifying the underlying binary search tree to store
sub-tree sizes will improve this to O(lg n), but short of implementing
my own red-black tree or AVL tree, this commit uses an array of offsets
to solve the problem.

The approach is naive because get_offset() is now O(n). A better
approach would be to use a cunning data structure to calculate prefix
sums in O(lg n).